renamed habitStatus to habitStatus-enum
